requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Mail==0.10.0
Flask-Limiter==3.8.0
Flask-Caching==2.3.0
Flask-Admin==1.6.1
python-dotenv==1.0.1
psycopg2-binary==2.9.10
gunicorn==23.0.0
requests==2.32.3
Authlib==1.3.2
redis==5.0.8
celery==5.4.0
Werkzeug==3.1.3
email-validator==2.2.0
itsdangerous==2.2.0

reportlab==4.4.3
